add types cohesionCenter et pointDeRassement
