+++ packages/lib/src/mongoSchema/index.ts
@@ -2,3 +2,5 @@ export type InterfaceExtended<T> = T & { _id: string; _doc?: T };
 
 export * from "./cle/classe";
 export * from "./cle/etablissement";
+export * from "./cohesionCenter";
+export * from "./pointDeRassemblement";
